+++ src/vietfin/providers/ssi/utils/etf_search.py
@@ -1,6 +1,6 @@
 """SSI Etf Search function."""
 
-import requests
+import httpx as requests
 
 from vietfin.providers.ssi.utils.helpers import ssi_headers
 from vietfin.abstract.vfobject import VfObject