+++ src/vietfin/providers/ssi/utils/index_constituents.py
@@ -1,6 +1,6 @@
 """SSI Index Constituents function."""
 
-import requests
+import httpx as requests
 
 from vietfin.providers.ssi.utils.helpers import ssi_headers
 from vietfin.abstract.vfobject import VfObject