+++ src/vietfin/providers/tcbs/utils/equity_fundamental_dividends.py
@@ -1,6 +1,6 @@
 """TCBS Equity Fundamental Dividends command."""
 
-import requests
+import httpx as requests
 
 from vietfin.abstract.vfobject import VfObject
 from vietfin.utils.helpers import (