+++ src/vietfin/providers/tcbs/utils/equity_price_quote.py
@@ -1,6 +1,6 @@
 """TCBS Equity Price Quote command."""
 
-import requests
+import httpx as requests
 
 from vietfin.abstract.vfobject import VfObject
 from vietfin.utils.helpers import (