+++ src/vietfin/providers/vndirect/utils/equity_discovery.py
@@ -1,6 +1,6 @@
 """VNDIRECT Equity Discovery group of functions."""
 
-import requests
+import httpx as requests
 
 from vietfin.abstract.vfobject import VfObject
 from vietfin.utils.helpers import (