+++ CHANGELOG.md
@@ -11,9 +11,9 @@
 - Add unit tests for this new features
 - Replace `datetime.utcfromtimestamp` with `datetime.fromtimestamp` since the previous will be [deprecated since python v3.12](https://docs.python.org/3/library/datetime.html#datetime.datetime.utcfromtimestamp)
 - Replace package [bs4](https://pypi.org/project/beautifulsoup4/) with [selectolax](https://github.com/rushter/selectolax) for better performance.
+- Replace package [requests](https://github.com/psf/requests) with [httpx](https://www.python-httpx.org/) for better performance.
 
 ongoing
-- Replace package [requests](https://github.com/psf/requests) with [httpx](https://www.python-httpx.org/) for better performance.
 - Add conditional import for playwright to retrieve cookies automatically when user wants to use `derivatives.futures.quote()` with provider `vdsc`.
 
 ## v0.1.0 (2024-03-08)